apps/requirements.txt
# LLM client (OpenAI-compatible, works with Ollama)
openai>=1.30.0

# Langfuse — LLM tracing and evals
langfuse>=4.0.0

# Phoenix (Docker): export traces via OTel + OpenInference below — not the arize-phoenix PyPI app server package
openinference-instrumentation-openai>=0.1.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0

# Prometheus metrics
prometheus-client>=0.20.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0       # Pretty console output for the lab demos
httpx>=0.27.0
